+++ ydb/tools/mnc/lib/legacy_commands/__init__.py
@@ -1,10 +1,8 @@
 from . import deploy
-from . import init
 from . import service
 
 
 modules = [
     deploy,
-    init,
     service,
 ]
